ForEachListApp and ListModificationChallenge done.
The first commit about the exercises to practice C# and their grammar and sintaxis language.
The first exercise show how create a List and loop it for show their contents.
The second one shows how you can insert a new object in a list in three diferents positions(first,last and middle),
both in a new list or existing list.